+++ app/src/main/java/org/example/app/system/SlashCommandRegister.java
@@ -1,4 +1,4 @@
-package org.example.app.listenner;
+package org.example.app.system;
 
 import java.util.List;
 import lombok.RequiredArgsConstructor;
