+++ app/src/main/resources/META-INF/org/example/app/application/example/version/VersionDao/getVersion.sql
